+++ src/main/kotlin/com/ebolo/studentmanager/views/teachers/SMTeacherTableFragment.kt
@@ -6,6 +6,7 @@ import com.ebolo.studentmanager.services.SMServiceCentral
 import com.ebolo.studentmanager.services.SMTeacherRefreshEvent
 import com.ebolo.studentmanager.services.SMTeacherRefreshRequest
 import com.ebolo.studentmanager.utils.SMCRUDUtils
+import com.ebolo.studentmanager.views.utils.ui.SMConfirmDialog
 import com.jfoenix.controls.JFXButton
 import com.jfoenix.controls.JFXTextField
 import javafx.collections.FXCollections
@@ -115,10 +116,15 @@ class SMTeacherTableFragment : Fragment() {
                     }
 
                     item("Xóa").action {
-                        serviceCentral.teacherService.deleteTeachers(selectionModel.selectedItems.map { it.id }.toList())
-                        fire(SMDataProcessRequest {
-                            fire(SMTeacherRefreshRequest)
-                        })
+                        find<SMConfirmDialog>(
+                            "dialogContent" to "Tiếp tục xóa?",
+                            "onOKClicked" to {
+                                serviceCentral.teacherService.deleteTeachers(selectionModel.selectedItems.map { it.id }.toList())
+                                fire(SMDataProcessRequest {
+                                    fire(SMTeacherRefreshRequest)
+                                })
+                            }
+                        ).openModal()
                     }
                 }
